In the start TCP, UDP and Seriel server, we are running a quick check (#242)
to see weather the thread in which the helpers are executed are the
Main thread or the spawned thread.

This infomation is passed to reactor.run() to the flag
installSignalHandlers. True is passed if the thread is a main
thread, else false.

This is done to prevent errors when running the helpers from a
spawned thread.